The UBO type can be moved out of the engine
